+++ dll/echarts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.3](https://git.easyops.local/anyclouds/next-core/compare/@dll/echarts@2.0.2...@dll/echarts@2.0.3) (2020-10-28)
+
+**Note:** Version bump only for package @dll/echarts
+
+## [2.0.2](https://git.easyops.local/anyclouds/next-core/compare/@dll/echarts@2.0.1...@dll/echarts@2.0.2) (2020-10-27)
+
+**Note:** Version bump only for package @dll/echarts
+
 ## [2.0.1](https://git.easyops.local/anyclouds/next-core/compare/@dll/echarts@2.0.0...@dll/echarts@2.0.1) (2020-10-15)
 
 **Note:** Version bump only for package @dll/echarts