+++ packages/build-config-factory/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.14](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@2.0.13...@easyops/build-config-factory@2.0.14) (2020-10-27)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+## [2.0.13](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@2.0.12...@easyops/build-config-factory@2.0.13) (2020-10-26)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
 ## [2.0.12](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@2.0.11...@easyops/build-config-factory@2.0.12) (2020-10-20)
 
 **Note:** Version bump only for package @easyops/build-config-factory