+++ packages/create-next-repo/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.33](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.32...@easyops/create-next-repo@1.0.33) (2020-10-28)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.32](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.31...@easyops/create-next-repo@1.0.32) (2020-10-27)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.31](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.30...@easyops/create-next-repo@1.0.31) (2020-10-26)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.30](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.29...@easyops/create-next-repo@1.0.30) (2020-10-26)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
 ## [1.0.29](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.28...@easyops/create-next-repo@1.0.29) (2020-10-22)
 
 **Note:** Version bump only for package @easyops/create-next-repo