+++ packages/webpack-config-factory/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/webpack-config-factory@2.1.0...@easyops/webpack-config-factory@2.1.1) (2020-10-28)
+
+**Note:** Version bump only for package @easyops/webpack-config-factory
+
+# [2.1.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/webpack-config-factory@2.0.5...@easyops/webpack-config-factory@2.1.0) (2020-10-27)
+
+### Features
+
+- **theme:** support dark theme ([fccc811](https://git.easyops.local/anyclouds/next-core/commits/fccc811))
+
+## [2.0.5](https://git.easyops.local/anyclouds/next-core/compare/@easyops/webpack-config-factory@2.0.4...@easyops/webpack-config-factory@2.0.5) (2020-10-26)
+
+**Note:** Version bump only for package @easyops/webpack-config-factory
+
 ## [2.0.4](https://git.easyops.local/anyclouds/next-core/compare/@easyops/webpack-config-factory@2.0.3...@easyops/webpack-config-factory@2.0.4) (2020-10-19)
 
 ### Bug Fixes